agent/ai-researcher/pyproject.toml
[project]
name = "ai-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-google-genai>=2.1.9",
    "langgraph>=0.6.3",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.48.0",
    "fastapi>=0.118.0",
    "uvicorn[standard]>=0.37.0",
    "supabase>=2.9.1",
    "grpcio>=1.75.1",
    "grpcio-tools>=1.75.1",
]
